+++ lib/animated_word_counter.dart
@@ -1,5 +1,6 @@
 library animated_word_counter;
 
 export 'src/animated_word_counter_widget.dart';
+export 'src/improved_animated_flip_counter.dart';
 export 'src/language_aware_word_counter.dart';
 export 'src/word_counter_languages.dart';